+++ mixed_congruential.h
@@ -2,11 +2,11 @@
 #include "linear_congruential.h"
 
 class MixedCongruentialGenerator : public LinearCongruentialGenerator {
-	int multiplier;
-	int increment;
+	long long multiplier;
+	long long increment;
 
 	void updateState() override;
 
 public:
-	MixedCongruentialGenerator(long long modulus, int multiplier, int increment);
+	MixedCongruentialGenerator(long long modulus, long long multiplier, long long increment);
 };
