+++ multiplicative_congruential.cpp
@@ -1,7 +1,7 @@
 #include "linear_congruential.h"
 #include "multiplicative_congruential.h"
 
-MultiplicativeCongruentialGenerator::MultiplicativeCongruentialGenerator(long long modulus, int multiplier) :
+MultiplicativeCongruentialGenerator::MultiplicativeCongruentialGenerator(long long modulus, long long multiplier) :
 	LinearCongruentialGenerator(modulus, 1), multiplier(multiplier) {};
 
 void MultiplicativeCongruentialGenerator::updateState() {
