+++ multiplicative_congruential.h
@@ -2,10 +2,10 @@
 #include "linear_congruential.h"
 
 class MultiplicativeCongruentialGenerator : public LinearCongruentialGenerator {
-	int multiplier;
+	long long multiplier;
 
 	void updateState() override;
 
 public:
-	MultiplicativeCongruentialGenerator(long long modulus, int multiplier);
+	MultiplicativeCongruentialGenerator(long long modulus, long long multiplier);
 };
